agents/requirements.txt
crewai>=0.11.0
litellm>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.0
